+++ src/page/index.js
@@ -1,3 +1,5 @@
 // @flow
-export { CommonTabbedPage } from "./common-tabbed-page";
-export { CommonPage } from "./common-page";
+export { CommonTabbedPage } from "./page-v1/common-tabbed-page";
+export { CommonPage } from "./page-v1/common-page";
+export { CommonTabbedPageV2 } from "./page-v2/common-tabbed-page-v2";
+export { CommonPageV2 } from "./page-v2/common-page-v2";