+++ src/page/page-v1/common-page.spec.js
@@ -1,9 +1,9 @@
 // @flow
 import "jest-dom/extend-expect";
 import React from "react";
-import { Alert, CommonSwitch } from "../index";
+import { Alert, CommonSwitch } from "../../index";
 import { CommonPage } from "./common-page";
-import { TestWrapper } from "../test-utils";
+import { TestWrapper } from "../../test-utils";
 import { cleanup, fireEvent, render } from "@testing-library/react";
 
 afterEach(cleanup);